requirements.txt
google-cloud-language
flask
numpy
Pillow
opencv-python